Move enums into their own files and namespace.
