+++ src/Domain/INSS.Platform.Portal.Domain/Enums/PaymentFrequency.cs
@@ -1,0 +1,10 @@
+﻿namespace INSS.Platform.Portal.Domain.Enums;
+
+public enum PaymentFrequency
+{
+    Hourly,
+    Daily,
+    Weekly,
+    Monthly,
+    Annually
+}
