+++ src/Presentation/INSS.Platform.Portal.Web.Components/Views/_ViewImports.cshtml
@@ -1,4 +1,5 @@
 ﻿@using INSS.Platform.Portal.Domain
+@using INSS.Platform.Portal.Domain.Enums
 @using INSS.Platform.Shared.Web.Utilities
 @using GovUk.Frontend.AspNetCore
